+++ python/boomer/common/statistics.pxd
@@ -37,15 +37,15 @@ cdef extern from "cpp/statistics.h" nogil:
 
         void resetSampledStatistics()
 
-        void addSampledStatistic(intp statisticIndex, uint32 weight)
+        void addSampledStatistic(uint32 statisticIndex, uint32 weight)
 
         void resetCoveredStatistics()
 
-        void updateCoveredStatistic(intp statisticIndex, uint32 weight, bool remove)
+        void updateCoveredStatistic(uint32 statisticIndex, uint32 weight, bool remove)
 
         AbstractRefinementSearch* beginSearch(uint32 numLabelIndices, const uint32* labelIndices)
 
-        void applyPrediction(intp statisticIndex, Prediction* prediction)
+        void applyPrediction(uint32 statisticIndex, Prediction* prediction)
 
 
 cdef class StatisticsProvider:
